// src/generated/console/requests.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  /** Represents non-fractional signed whole numeric values. */
  BigInt: { input: any; output: any };
  /** ISO8601 formatted date (yyyy-mm-dd) */
  Date: { input: any; output: any };
  DateTime: { input: any; output: any };
  Decimal: { input: any; output: any };
  EmailAddress: { input: any; output: any };
  IpAddress: { input: any; output: any };
  /** Represents untyped JSON */
  JSON: { input: any; output: any };
  PhoneNumber: { input: any; output: any };
  URL: { input: any; output: any };
};
// ...
export type AdminUser = {
  __typename: 'AdminUser';
  createdAt: Scalars['DateTime']['output'];
  email: Scalars['EmailAddress']['output'];
  id: Scalars['ID']['output'];
  roles?: Maybe<Array<Scalars['String']['output']>>;
  updatedAt: Scalars['DateTime']['output'];
};
// ...
export type Client = {
  __typename: 'Client';
  audits?: Maybe<AuditConnection>;
  businessStructure?: Maybe<Scalars['String']['output']>;
  canArchive: Scalars['Boolean']['output'];
  cbsEnabled: Scalars['Boolean']['output'];
  companyNumber?: Maybe<Scalars['String']['output']>;
  contacts: ContactConnection;
  createdAt: Scalars['DateTime']['output'];
  dashboardUri?: Maybe<Scalars['String']['output']>;
  deletedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Reference number for the client (eg. CLI-0001) */
  displayReferenceNumber: Scalars['String']['output'];
  /** @deprecated emailAddress is no longer available on the client. Use the default contact email instead. */
  emailAddress?: Maybe<Scalars['EmailAddress']['output']>;
  engagements?: Maybe<EngagementConnection>;
  faxNumber?: Maybe<Scalars['PhoneNumber']['output']>;
  gstRegistered?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  invoices?: Maybe<ClientBillingInvoiceConnection>;
  /** Is revenue from this client excluded from the dashboard and other reports? */
  isExcludedFromReports: Scalars['Boolean']['output'];
  /** Whether the practice applies surcharges to pass on payment fees to this client */
  isSurchargeEnabled: Scalars['Boolean']['output'];
  name: Scalars['String']['output'];
  notes?: Maybe<Scalars['String']['output']>;
  notificationsSettings: NotificationsSettingsConnection;
  paymentMethods?: Maybe<Array<PaymentMethod>>;
  phoneNumber?: Maybe<Scalars['PhoneNumber']['output']>;
  /** Indicates if the client is a placeholder client for public previews */
  placeholder: Scalars['Boolean']['output'];
  practice: Practice;
  proposals?: Maybe<ProposalConnection>;
  recipientGroup: RecipientGroupConnection;
  state: ClientState;
  type: ClientType;
  updatedAt: Scalars['DateTime']['output'];
  websiteUrl?: Maybe<Scalars['URL']['output']>;
};
// ...
export type ClientConnection = {
  __typename: 'ClientConnection';
  /** A list of edges. */
  edges: Array<ClientEdge>;
  /** A list of nodes. */
  nodes: Array<Client>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type CustomTemplateConnection = {
  __typename: 'CustomTemplateConnection';
  /** A list of edges. */
  edges: Array<CustomTemplateEdge>;
  /** A list of nodes. */
  nodes: Array<ProposalEditorCustomTemplate>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  templatesYaml: Scalars['String']['output'];
};
// ...
export enum PaymentsOnboardedPresetType {
  BANK_ACCOUNT_PAYMENT_DISCOUNT = 'BANK_ACCOUNT_PAYMENT_DISCOUNT',
  HIGH_VALUE = 'HIGH_VALUE',
  INTERNAL = 'INTERNAL',
  PARTNERSHIP_CARBON_GROUP = 'PARTNERSHIP_CARBON_GROUP',
  PARTNERSHIP_TAX_ASSIST = 'PARTNERSHIP_TAX_ASSIST',
  PARTNERSHIP_XEINADIN_GROUP = 'PARTNERSHIP_XEINADIN_GROUP',
  SUPER_HIGH_VALUE = 'SUPER_HIGH_VALUE',
  TWENTY_TWENTY_THREE_NO_FEES_THREE_MONTHS = 'TWENTY_TWENTY_THREE_NO_FEES_THREE_MONTHS',
}
// ...
export type PaymentsSettings = {
  __typename: 'PaymentsSettings';
  audits?: Maybe<AuditConnection>;
  /** Whether this practice is collecting payments via credit card */
  creditCardAccepted: Scalars['Boolean']['output'];
  /** Whether this practice is collecting payments via direct debit */
  directDebitAccepted: Scalars['Boolean']['output'];
  /** Whether direct debit collections can be turned on for this practice */
  directDebitSupported: Scalars['Boolean']['output'];
  /** Scheduled date disbursals will be disabled for this practice. Clears once a practice's disbursals have been disabled */
  disableDisbursalsAt?: Maybe<Scalars['DateTime']['output']>;
  /** Disbursals Disablement Linked Payment Collection */
  disbursalsDisablementLinkedCollection?: Maybe<PaymentsCollection>;
  /** Disbursals Disablement Type */
  disbursalsDisablementType?: Maybe<DisbursalsDisablementType>;
  /** Fraud Payment Threshold */
  fraudPaymentThreshold?: Maybe<Money>;
  hasDisbursalAccount: Scalars['Boolean']['output'];
  /** Are collections enabled? */
  isCollectionsEnabled: Scalars['Boolean']['output'];
  /** Are disbursals enabled? */
  isDisbursalsEnabled: Scalars['Boolean']['output'];
  /** Whether surcharges are restricted to only the 'credit_card' surcharge category */
  isSurchargeCreditCardsOnly: Scalars['Boolean']['output'];
  /** Whether a practice can use surcharges to pass on payment fees to their clients */
  isSurchargeEnabled: Scalars['Boolean']['output'];
  /** Settings configuration applied to practice */
  onboardedPreset?: Maybe<PaymentsOnboardedPresetType>;
  /** Number of days a payment will be collected after an invoice for one-off payment is raised */
  onceOffTerms: Scalars['Int']['output'];
  /** The percentage fee discount applied to any payment fees incurred */
  paymentFeeDiscount?: Maybe<Percentage>;
  /** Whether this practice requires a valid payment method from a client when a proposal is accepted */
  paymentMethodRequired: Scalars['Boolean']['output'];
  /** Number of days a payment will be collected after an invoice for a recurring payment is raised */
  recurringTerms: Scalars['Int']['output'];
  /** Description of the most recent change to surcharge settings */
  surchargeLastUpdateMessage?: Maybe<Scalars['String']['output']>;
  /** Rate at which applicable cards are surcharged */
  surchargeRate: Percentage;
};
// ...
export type Plan = {
  __typename: 'Plan';
  billingPeriod: PlanBillingPeriod;
  code: Scalars['String']['output'];
  /** If set, defines the active client limit for this practice overriding the default provided by their plan. */
  customActiveClientLimit?: Maybe<Scalars['Int']['output']>;
  isEligible: Scalars['Boolean']['output'];
  isLegacy: Scalars['Boolean']['output'];
  maxClients: Scalars['Int']['output'];
  name: Scalars['String']['output'];
  overagesEnabled: Scalars['Boolean']['output'];
  price: Money;
  tier: PlanNameEnum;
};
// ...
export type Practice = {
  __typename: 'Practice';
  audits?: Maybe<AuditConnection>;
  billingCurrency: Currency;
  brandingTheme: BrandingTheme;
  bulkKickOffs?: Maybe<BulkKickOffConnection>;
  client?: Maybe<Client>;
  clients?: Maybe<ClientConnection>;
  /** List of revenue brackets for cmi questions based on the current practice region */
  cmiRevenueBrackets: Array<RevenueBracket>;
  countryCode?: Maybe<Scalars['String']['output']>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  currencySymbol: Scalars['String']['output'];
  customTemplates?: Maybe<CustomTemplateConnection>;
  customerSuccessMaturityIndex?: Maybe<CustomerSuccessMaturityIndexUnion>;
  customerSuccessMaturityIndices?: Maybe<CustomerSuccessMaturityIndexUnionConnection>;
  /** Default service grouping for new proposals. This will be used unless otherwise changed within the proposal editor. */
  defaultBillingGroupedBy: BillingGroupedByType;
  disabled: Scalars['Boolean']['output'];
  disabledReasonCategory?: Maybe<PracticeDisabledReasonCategory>;
  email?: Maybe<Scalars['EmailAddress']['output']>;
  engagement?: Maybe<Engagement>;
  engagementLetters?: Maybe<CustomerSuccessEngagementLetterConnection>;
  engagements?: Maybe<EngagementConnection>;
  feature?: Maybe<Feature>;
  featureFlag: FeatureFlag;
  features: Array<Feature>;
  id: Scalars['ID']['output'];
  integrations?: Maybe<IntegrationConnection>;
  intent: PracticeIntentEnum;
  interchangeFees: InterchangeFeeConnection;
  internal: Scalars['Boolean']['output'];
  ledgerName?: Maybe<Scalars['String']['output']>;
  locale?: Maybe<Scalars['String']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  notificationsSettings: NotificationsSettingsConnection;
  /** Practice assigned payment gateways */
  paymentGateways?: Maybe<Array<PracticePaymentGateway>>;
  paymentMethodFees?: Maybe<PaymentMethodFeeConnection>;
  paymentSettings: PaymentsSettings;
  paymentsDisbursals?: Maybe<PaymentsDisbursalConnection>;
  phone?: Maybe<Scalars['String']['output']>;
  plan?: Maybe<Plan>;
  preferredProposalEditor?: Maybe<ProposalEditor>;
  principal?: Maybe<User>;
  proposal?: Maybe<Proposal>;
  proposalTemplatePublishRegion?: Maybe<Scalars['String']['output']>;
  /** List of Quickbooks Recurring Transactions */
  quickbooksRecurringTransactions?: Maybe<QuickbooksRecurringTransactionConnection>;
  recurlyAccountUrl?: Maybe<Scalars['URL']['output']>;
  referenceNumber: Scalars['ID']['output'];
  /**
   * The referral copy published for a particular region.
   *                       Only applicable on special practices with 'proposal_template_publish_region' defined.
   */
  referralCopy?: Maybe<CustomerSuccessReferralCopy>;
  region?: Maybe<Scalars['String']['output']>;
  renewalsImports?: Maybe<RenewalsImportConnection>;
  services: ServiceTypeConnection;
  slug: Scalars['String']['output'];
  stripeIntegrationAccount?: Maybe<StripeIntegrationAccount>;
  templates?: Maybe<TemplateConnection>;
  tfaEffectiveFrom?: Maybe<Scalars['Date']['output']>;
  timeZone: TimeZone;
  trialEndDate?: Maybe<Scalars['Date']['output']>;
  twoFactorAuthRequired: Scalars['Boolean']['output'];
  /** Find a User by ID */
  user?: Maybe<User>;
  users?: Maybe<UserConnection>;
  /** List of Xero Repeating Invoices */
  xeroRepeatingInvoices?: Maybe<XeroIntegrationRepeatingInvoiceConnection>;
};
// ...
export enum ProposalEditor {
  CPE = 'CPE',
  NPE = 'NPE',
}
// ...
export type Query = {
  __typename: 'Query';
  client?: Maybe<Client>;
  clients?: Maybe<ClientConnection>;
  codeVersion: Scalars['String']['output'];
  currentUser: AdminUser;
  customTemplate?: Maybe<ProposalEditorCustomTemplate>;
  customTemplates?: Maybe<CustomTemplateConnection>;
  engagement?: Maybe<Engagement>;
  environment: Scalars['String']['output'];
  /** Client Billing Invoice */
  invoice?: Maybe<ClientBillingInvoice>;
  paymentsDisbursal?: Maybe<PaymentsDisbursal>;
  paymentsPayment?: Maybe<PaymentsPayment>;
  paymentsPaymentCollection?: Maybe<PaymentsCollection>;
  paymentsRefund?: Maybe<PaymentsRefund>;
  payout?: Maybe<StripePayout>;
  practice: Practice;
  practices: PracticeConnection;
  /** Returns a Ignition object based on a related external Stripe object ID */
  stripeIdSearch?: Maybe<StripeIdSearchUnionConnection>;
  /** Find a User by UUID */
  user?: Maybe<User>;
  users: UserConnection;
};
// ...
export type StripeIntegrationAccount = {
  __typename: 'StripeIntegrationAccount';
  bacsDebitSun?: Maybe<BacsDebitSunType>;
  dashboardUri: Scalars['URL']['output'];
  id: Scalars['ID']['output'];
  stripeId: Scalars['String']['output'];
};
// ...
export type PracticeQuery = {
  __typename: 'Query';
  practice: {
    __typename: 'Practice';
    id: string;
    referenceNumber: string;
    paymentSettings: {
      __typename: 'PaymentsSettings';
      isCollectionsEnabled: boolean;
      isDisbursalsEnabled: boolean;
      onboardedPreset?: Types.PaymentsOnboardedPresetType | null;
    };
    plan?: { __typename: 'Plan'; name: string } | null;
    stripeIntegrationAccount?: {
      __typename: 'StripeIntegrationAccount';
      dashboardUri: any;
    } | null;
  };
};